requirements.txt
django==3.2
Pillow
psycopg2
dj-database-url
python-decouple
gunicorn
whitenoise
#IMAGENES
cloudinary
#django-paypal
djangorestframework
django-cors-headers